requirements.txt
Flask==3.0.3
Flask-Caching==2.3.0
PyTesseract==0.3.10
EasyOCR==1.7.1
PaddleOCR==2.8.1
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
Werkzeug==3.0.2